Move code from storybook to component's implementation, unify combobox under dropdowncomboboxmenu
